+++ internal/scaling/scaling.go
@@ -24,6 +24,7 @@ import (
 	"fmt"
 	"time"
 
+	argov1alpha1 "github.com/argoproj/argo-rollouts/pkg/apis/rollouts/v1alpha1"
 	"github.com/golang/glog"
 	"github.com/jthomperoo/custom-pod-autoscaler/v2/config"
 	"github.com/jthomperoo/custom-pod-autoscaler/v2/evaluate"
@@ -191,6 +192,8 @@ func (s *Scale) getReplicaCount(resource metav1.Object) (*int32, error) {
 	case *appsv1.StatefulSet:
 		return v.Spec.Replicas, nil
 	case *corev1.ReplicationController:
+		return v.Spec.Replicas, nil
+	case *argov1alpha1.Rollout:
 		return v.Spec.Replicas, nil
 	default:
 		return nil, fmt.Errorf("Unsupported resource of type %T", v)